requirements.txt
APScheduler==3.6.3
atomicwrites==1.4.0
attrs==21.4.0
backports.zoneinfo==0.2.1
black==22.3.0
cachetools==4.2.2
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
flake8==4.0.1
idna==3.3
importlib-metadata==4.2.0
iniconfig==1.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.1
pluggy==1.0.0
py==1.11.0
pycodestyle==2.8.0
pyflakes==2.4.0
pyparsing==3.0.8
pytest==7.1.1
python-dotenv==0.20.0
python-telegram-bot==13.11
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
six==1.16.0
tomli==2.0.1
tornado==6.1
typed-ast==1.5.2
typing-extensions==4.1.1
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
zipp==3.8.0
